split debugger files and add `reattachOwner`
